chore: Fix code style, change CategoryPopover
