+++ src/components/CategoryPopover/CategoryPopover.jsx
@@ -1,11 +1,33 @@
-import React from 'react'
+import React, { useState } from 'react'
 import Popover from '../Popover/Popover';
-import CategogyPopoverContent from './CategoryPopoverContent';
+import * as Styled from './styles'
 
-const CategoryPopover = () => (
-  <Popover popoverLabel='Category'>
-    <CategogyPopoverContent/>
-  </Popover>
-);
+const CategoryPopover = () => {
+  const categories = ['Shopping', 'Entertainment', 'Car', 'Bills', 'Food', 'Home', 'Education', 'Other'];
+  const [categoryValue, setValue] = useState('');
+
+  const setCategory = (category) => {
+    setValue(category);
+  };
+
+  return (
+    <Popover popoverLabel='Category'>
+      <>
+        <Styled.CategoryWrap>
+          {categories.map(category => (
+            <Styled.CategoryItem
+              key={category}
+              isActive={categoryValue === category}
+              onClick={() => setCategory(category)}
+            >
+              {category}
+            </Styled.CategoryItem>
+          ))}
+        </Styled.CategoryWrap>
+        <Styled.ButtonItem text='Apply' onClick={() => console.log(categoryValue)}/>
+      </>
+    </Popover>
+  )
+};
 
 export default CategoryPopover;
