--- src/components/CategoryPopover/CategoryPopoverContent/index.js
@@ -1,1 +1,0 @@
-export { default } from './CategoryPopoverContent';
